Added new classes and interfaces
