--- src/main/java/Human/Address/Address.java
@@ -1,4 +1,0 @@
-package Human.Address;
-
-public class Address {
-}
