+++ src/main/java/Lukyanov/HumanStorageImpl.java
@@ -1,0 +1,25 @@
+package Lukyanov;
+
+import java.util.List;
+
+public class HumanStorageImpl<Human> implements HumanStorage<Human> {
+    @Override
+    public Human getEntity(long id) {
+        return null;
+    }
+
+    @Override
+    public List<Human> getAllEntities() {
+        return null;
+    }
+
+    @Override
+    public void saveEntity(Human human) {
+
+    }
+
+    @Override
+    public void saveAllEntities(List<Human> humans) {
+
+    }
+}
