--- src/main/java/MyInterface.java
@@ -1,8 +1,0 @@
-import java.util.List;
-
-public interface MyInterface<T> {
-    T getEntity();
-    List<T> getAllEntities();
-    void saveEntity();
-    void saveAllEntities();
-}
